getAll ok, with pagination  using DataCollection class

+++ Models/ProfileDancer.cs
@@ -5,7 +5,7 @@
         public int ProfileDanceId { get; set; }
         public int DanceLevel_FK { get; set; }
         public int DanceRol_FK { get; set; }
-        public decimal? Height { get; set; }
+        public double? Height { get; set; }
         public DanceRol? DanceRol { get; }
         public DanceLevel? DanceLevel { get; }
