+++ ServiceEventHandler/CreateCommand/ProfileDancerCreateCommand.cs
@@ -8,7 +8,7 @@ namespace ServiceEventHandler.CreateCommand
     {
         public int DanceLevel_FK { get; set; }
         public int DanceRol_FK { get; set; }
-        public decimal Height { get; set; }
+        public double Height { get; set; }
 
     }
 }
